Extract random imagesource
