+++ volumina/pixelpipeline/imagesources/__init__.py
@@ -38,6 +38,7 @@ from .grayscale import GrayscaleImageSource
 from .alphamodulated import AlphaModulatedImageSource
 from .colortable import ColortableImageSource
 from .rgba import RGBAImageSource
+from .random import RandomImageSource
 
 _has_vigra = True
 try:
@@ -47,27 +48,6 @@ except ImportError:
 
 
 logger = logging.getLogger(__name__)
-
-class RandomImageSource(ImageSource):
-    """Random noise image for testing and debugging."""
-
-    def request(self, qrect, along_through=None):
-        assert isinstance(qrect, QRect)
-        s = rect2slicing(qrect)
-        shape = slicing2shape(s)
-        return RandomImageRequest(shape)
-
-
-class RandomImageRequest(RequestABC):
-    def __init__(self, shape):
-        self.shape = shape
-
-    def wait(self):
-        d = (np.random.random(self.shape) * 255).astype(np.uint8)
-        assert d.ndim == 2
-        img = gray2qimage(d)
-        return img.convertToFormat(QImage.Format_ARGB32_Premultiplied)
-
 
 ##
 ## Sources that produce QGraphicsItems isntead of QImages
